% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colname_from_colnumber}
\alias{colname_from_colnumber}
\title{Spreadsheet column name from number}
\usage{
colname_from_colnumber(colnumber)
}
\arguments{
\item{colnumber}{An integer. The number of the column.}
}
\description{
Convert an integer number to the equivalent column name in a spreadsheet. This is an helper function used internaly in \code{\link{descvars_skeleton}} to generate the variable column. For example, column 28 is "AB".
}
\note{
This function have a bug : \code{colname_from_colnumber(53)} give AA instead of BA. The bug is \href{https://github.com/jomuller/vartors/issues/59}{listed in the bugtracking system as the issue 59} and will be fixed in a further version.
}
\examples{
colname_from_colnumber(3)
# "C"
colname_from_colnumber(27)
# "AA"
colname_from_colnumber(666)
# "XP"
}
\author{
Joris Muller
}
\keyword{internal}

