% Generated by roxygen2 (4.0.1): do not edit by hand
\name{database_def}
\alias{database_def}
\title{Create a DatabaseDef object}
\usage{
database_def(DatabaseDef)
}
\arguments{
\item{DatabaseDef}{A list containing \code{VariableDef} objects}
}
\description{
Constructor of the class DatabaseDef. Used only internaly for the moment.
}
\keyword{internal}

