% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{DatabaseDef-class}
\alias{DatabaseDef-class}
\title{Class DatabaseDef}
\description{
The \code{DatabaseDef} class is used to store properly the definition of several variables. It is created by \link{import_vardef}.
}
\section{Slots}{

\describe{
\item{\code{DatabaseDef}}{A list containing \code{VariableDef} objects.}
}}
\note{
For the moment, this class is only a convenient way to store a list of \code{\link{VariableDef-class}} objects. This class will be extended in a future version of \pkg{vartors} to add the pathfile, a global description and others informations about the database.
}
\examples{
# Create a DatabaseDef from a definition of variable table
suppressWarnings(
a_DatabaseDef_object <- import_vardef(sample_descvar)
)

# Show the list of the definition of variable
a_DatabaseDef_object

# Check the class
class(a_DatabaseDef_object)
}
\author{
Joris Muller
}
\seealso{
\code{DatabaseDef} objects are created by \link{import_vardef} function for the moment. It store \code{link[=VariableDef-class]{VariableDef}} objects in a list. To create a single definition of variable, use \code{\link{vardef}}.
}
\keyword{classes}

