% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{example_df}
\alias{example_df}
\title{Sample data.frame to test vartors}
\format{A data frame with 100 rows and 10 variables}
\description{
A small data set randomly generated to simulate 100 observation on an hypothetical survey with 10 columns. Include NA's.

The variables are :
\itemize{
    \item{\code{subject}}{ A numeric. Unique number of the subject.}
\item{\code{initial}}{ A character. Initials of the subject.}
\item{\code{birth}}{ Birthdate}
\item{\code{sex}}{ A factor with levels \code{male} \code{female}}
\item{\code{study_level}}{ An ordered factor with levels \code{primary} < \code{secondary} < \code{superior}}
\item{\code{heigh}}{ A numeric}
\item{\code{weight}}{ A numeric}
\item{\code{siblings}}{ A numeric}
\item{\code{Q1}}{ A numeric : question 1}
\item{\code{Q2}}{ A numeric : question 2}
}
}
\seealso{
\link{bad_database}, a version closer to a real-life database.
}
\keyword{datasets}

