% Generated by roxygen2 (4.0.1): do not edit by hand
\name{excel_skeleton}
\alias{excel_skeleton}
\title{Write an Excel VariableDef skeleton}
\usage{
excel_skeleton(filepath = "variables_description.xlsx")
}
\arguments{
\item{filepath}{path to the file to create}
}
\description{
Write an Excel VariableDef skeleton
}

