% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{$,ScriptOutput-method}
\alias{$,ScriptOutput-method}
\alias{$,VariableDef-method}
\alias{$<-,ScriptOutput-method}
\title{Get value of a ScriptOutput object}
\usage{
\S4method{$}{ScriptOutput}(x, name)

\S4method{$}{ScriptOutput}(x, name) <- value

\S4method{$}{VariableDef}(x, name)
}
\arguments{
\item{x}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{name}{
    A literal character string or a \link{name} (possibly \link{backtick}
    quoted).  For extraction, this is normally (see under
    \sQuote{Environments}) partially matched to the \code{\link{names}}
    of the object.
  }

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
Get value of a ScriptOutput object

Replace names of ScriptOutput

Get value for VariableDef object
}
\keyword{internal}

