% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_vartors_template}
\alias{is_vartors_template}
\title{Detect if a specified file is a vartors template}
\usage{
is_vartors_template(filepath, lines_to_read = 5)
}
\arguments{
\item{filepath}{Path to the file}

\item{lines_to_read}{How many first lines to read to find the tag <vartors template>}
}
\description{
Detect if a specified file is a vartors template
}
\seealso{
\code{\link{list_templates}}
}
\keyword{internal}
\keyword{template}

