% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ScriptOutput-class}
\alias{ScriptOutput-class}
\title{Class ScriptOutput}
\description{
Class to store the script output
}
\section{Slots}{

\describe{
\item{\code{output}}{body of the script}

\item{\code{language}}{Language of the script. Is "R", "Rmd" or "Rnw"}

\item{\code{header}}{Header of the script}

\item{\code{footer}}{Footer of the script}
}}
\author{
Joris Muller
}
\seealso{
The constructor method is \code{\link{create_script}}. \code{ScriptOutput} objects have also a write method to create easily a file
}
\keyword{classes}

