% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ScriptTemplate-class}
\alias{ScriptTemplate-class}
\title{Class ScriptTemplate}
\description{
The \code{ScriptTemplate} class is used to store properly the script template. It consists of various blocs, each for each type (numeric, factor...)
}
\section{Slots}{

\describe{
\item{\code{language}}{A length-one character vector. The extension of the language used in the template. Should be \code{R}, \code{Rmd} or \code{Rnw}}

\item{\code{original_script}}{The original script}

\item{\code{blocs}}{A list. Each element of the list is a character vector with the lines for a type.}
}}
\author{
Joris Muller
}
\seealso{
The main function to construct a \code{ScriptTemplate} object is \code{\link{import_template}}. The constructor is \code{\link{script_template}}. More information about template in the \link[=ScriptTemplate]{dedicated documentation}.
}
\keyword{classes}
\keyword{template}

