% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vardef}
\alias{vardef}
\title{Create a VariableDef object}
\usage{
vardef(varlabel, description, rname, type = "not_used", comment, unit,
  levels = NULL, names = NULL)
}
\arguments{
\item{varlabel}{A character. Used to label properly the plots and tables in output.}

\item{description}{A character. A description of the variable}

\item{rname}{A character. It's the name of the variable used in R.}

\item{comment}{A character.}

\item{type}{A character. Must be one of the following : numeric, integer, factor, ordered, character, date or not_used.}

\item{unit}{A character. Could be used for the format of a date (by default aa/mm/yyyy).}

\item{levels}{A character vector. Describe the levels used for a vector.}

\item{names}{A character vector of the same size than number of levels or empty.}
}
\description{
Constructor of the \code{\link[=VariableDef-class]{VariableDef}} class. A \code{\link[=VariableDef-class]{VariableDef}} object stores all the data needed to process a variable in the package.
}
\author{
Joris Muller
}
\seealso{
\link{VariableDef-class} and \link{DatabaseDef-class}
}

